aula 10

+++ lib/cart/models/cart.dart
@@ -6,7 +6,7 @@ class Cart extends Equatable {
   const Cart({this.items = const <Item>[]});
   final List<Item> items;
 
-  int get totalPrice {
+  double get totalPrice {
     return items.fold(0, (total, current) => total + current.price);
   }
